backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0